phidata-agent-serverless-api-basic/lambdas/agents/requirements.txt
openai
boto3
s3fs
httpx
phidata
pinecone